Backend/gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-jose
prometheus-fastapi-instrumentator
pydantic
typing-extensions
python-multipart  
pytz